Fix createTestBitmap to create 1x1 images
Previously it didn't paint any color to the bitmap, so for images with width/height = 1 we end up with empty image. Then the testReadBitmap_1x1 test failed probably because Pix and Bitmap handled differently this blank image.

Note this change fixes testReadBitmap_1x1 test. Also it makes testWriteBitmap_1x1 to fail, but it works if we use Color.BLACK there. Most of the write tests are obviously broken.